+++ src/quacc/recipes/espresso/_base.py
@@ -13,6 +13,7 @@ from quacc.calculators.espresso.espresso import (
     EspressoProfile,
     EspressoTemplate,
 )
+from quacc.calculators.espresso.utils import remove_conflicting_kpts_kspacing
 from quacc.runners.ase import run_calc, run_opt
 from quacc.schemas.ase import summarize_opt_run, summarize_run
 from quacc.utils.dicts import recursive_dict_merge
@@ -24,7 +25,7 @@ if TYPE_CHECKING:
     from quacc.utils.files import Filenames, SourceDirectory
 
 
-def base_fn(
+def run_and_summarize(
     atoms: Atoms | None = None,
     preset: str | None = None,
     template: EspressoTemplate | None = None,
@@ -85,7 +86,7 @@ def base_fn(
     )
 
 
-def base_opt_fn(
+def run_and_summarize_opt(
     atoms: Atoms | None = None,
     preset: str | None = None,
     relax_cell: bool = False,
@@ -195,6 +196,8 @@ def _prepare_atoms(
         Atoms object with attached Espresso calculator.
     """
     atoms = Atoms() if atoms is None else atoms
+    calc_defaults = calc_defaults or {}
+    calc_swaps = calc_swaps or {}
 
     calc_defaults["input_data"] = Namelist(calc_defaults.get("input_data"))
     calc_swaps["input_data"] = Namelist(calc_swaps.get("input_data"))
@@ -204,6 +207,8 @@ def _prepare_atoms(
     if binary in ALL_KEYS:
         calc_defaults["input_data"].to_nested(binary=binary, **calc_defaults)
         calc_swaps["input_data"].to_nested(binary=binary, **calc_swaps)
+
+    calc_defaults = remove_conflicting_kpts_kspacing(calc_defaults, calc_swaps)
 
     calc_flags = recursive_dict_merge(calc_defaults, calc_swaps)