+++ tests/core/calculators/espresso/test_espresso.py
@@ -147,28 +147,21 @@ def test_espresso_kpts():
 def test_output_handler(tmp_path, monkeypatch):
     monkeypatch.chdir(tmp_path)
 
-    parameters = {"input_data": {"system": {"outdir": "../test3/test2/test1"}}}
+    parameters = {"input_data": {"control": {"outdir": "../test3/test2/test1"}}}
 
     test_path = "../test3/test2/test1"
     fake_template = EspressoTemplate()
-    with pytest.raises(ValueError, match="because it is"):
-        new_parameters = fake_template._output_handler(parameters, Path())
+    new_parameters = fake_template._output_handler(parameters, Path())
+
+    assert str(new_parameters["input_data"]["control"]["outdir"]) != test_path
 
     test_path = "/test3/test2/test1"
-    parameters["input_data"]["system"]["outdir"] = test_path
+    parameters["input_data"]["control"]["outdir"] = test_path
     fake_template = EspressoTemplate()
-    with pytest.raises(ValueError, match="because it is"):
-        new_parameters = fake_template._output_handler(parameters, Path())
 
-    test_path = Path("test3/test2/test1")
-    parameters["input_data"]["system"]["outdir"] = test_path
-    fake_template = EspressoTemplate()
     new_parameters = fake_template._output_handler(parameters, Path())
-    assert (
-        new_parameters["input_data"]["system"]["outdir"]
-        == test_path.expanduser().resolve()
-    )
-    assert Path(tmp_path, "test3/test2/test1").exists()
+
+    assert str(new_parameters["input_data"]["control"]["outdir"]) != test_path
 
 
 def test_bad_calculator_params():